Move resolver wrappers into sub-packages
