+++ pkg/resolvers/coalesced/coalesced_resolver.go
