+++ pkg/resolvers/coalesced/coalesced_resolver_test.go
